updated dep's and refactored code
